refactor : Product 의 Rate 관련 속성 int -> float 변경
- int 일때 매번 100를 나누어주는 번거로움도 있고 직관적이지 않다는 문제로 float로 변경
- 추후 다른 엔티티의 rate도 float로 변경

+++ src/main/java/com/freeder/buclserver/domain/product/entity/Product.java
@@ -80,17 +80,17 @@ public class Product extends TimestampMixin {
 	private TaxStatus taxStatus;
 
 	@Column(name = "margin_rate")
-	private int marginRate;
+	private float marginRate;
 	@Column(name = "tax_rate")
-	private int taxRate;
+	private float taxRate;
 	@Column(name = "discount_rate")
-	private int discountRate;
+	private float discountRate;
 
 	@Column(name = "consumer_reward_rate")
-	private int consumerRewardRate;
+	private float consumerRewardRate;
 
 	@Column(name = "business_reward_rate")
-	private int businessRewardRate;
+	private float businessRewardRate;
 
 	@Column(name = "image_path", length = 5000)
 	private String imagePath;
